requirements.txt
Django==1.6
Pillow==2.7.0
argparse==1.2.1
django-admin-bootstrapped==2.3.3
django-annoying==0.8.1
django-appconf==1.0.1
django-bootstrap-toolkit==2.15.0
django-crispy-forms==1.4.0
django-embed-video==0.11
django-endless-pagination==2.0
django-imagekit==3.2.6
django-jsonview==0.4.3
elasticsearch==1.4.0
pilkit==1.1.12
pytz==2014.10
requests==2.5.3
simplejson==3.6.5
six==1.9.0
urllib3==1.10.2
wsgiref==0.1.2
